merge+training draft
