+++ src/careamics/config/nm_model.py
@@ -1,8 +1,7 @@
 """Noise models config."""
 
 from pathlib import Path
-from pathlib import Path
-from typing import Any, Literal, Union, Union
+from typing import Any, Literal, Union
 
 import numpy as np
 from pydantic import BaseModel, ConfigDict, Field, model_validator
@@ -26,35 +25,22 @@ class GaussianMixtureNmModel(BaseModel):
     n_gaussian: int = Field(default=1, ge=1)
     """Number of gaussians in the mixture."""
 
-    n_coeff: int = Field(default=2, ge=2)
     """Number of coefficients to describe the functional relationship between gaussian
     parameters and the signal. 2 implies a linear relationship, 3 implies a quadratic
     relationship and so on."""
+    n_coeff: int = Field(default=2, ge=2)
 
-    min_signal: float = Field(default=0.0, ge=0.0)
     """Minimum signal intensity expected in the image."""
+    min_signal: float = Field(default=0.0, ge=0.0)
 
-    max_signal: float = Field(default=1.0, ge=0.0)
     """Maximum signal intensity expected in the image."""
+    max_signal: float = Field(default=1.0, ge=0.0)
 
-    min_sigma: Any = Field(default=200.0, ge=0.0)  # TODO took from nb in pn2v
     """All values of `standard deviation` below this are clamped to this value."""
+    min_sigma: Any = Field(default=200.0, ge=0.0)  # TODO took from nb in pn2v
 
-    tol: float = Field(default=1e-10)
     """Tolerance used in the computation of the noise model likelihood."""
-
-
-# The noise model is given by a set of GMMs, one for each target
-# e.g., 2 target channels, 2 noise models
-class MultiChannelNmModel(BaseModel):
-    """Noise Model that aggregates the noise models for single channels."""
-
-    # TODO: check that this model config is OK
-    model_config = ConfigDict(
-        validate_assignment=True, arbitrary_types_allowed=True, extra="allow"
-    )
-    noise_models: list[GaussianMixtureNmModel]
-    """List of noise models, one for each target channel."""
+    tol: float = Field(default=1e-10)
 
     @model_validator(mode="after")  # TODO isn't it the best ever fucntion name ? :)
     def validate_path_to_pretrained_vs_data_to_train(self: Self):
@@ -79,3 +65,17 @@ class MultiChannelNmModel(BaseModel):
                 "provided or both signal and observation in form of paths"
                 "or numpy arrays"
             )
+
+
+# The noise model is given by a set of GMMs, one for each target
+# e.g., 2 target channels, 2 noise models
+class MultiChannelNmModel(BaseModel):
+    """Noise Model that aggregates the noise models for single channels."""
+
+    # TODO: check that this model config is OK
+    model_config = ConfigDict(
+        validate_assignment=True, arbitrary_types_allowed=True, extra="allow"
+    )
+    """List of noise models, one for each target channel."""
+    noise_models: list[GaussianMixtureNmModel]
+
